+++ runner/src/server/plugins/engine/components/ConditionalFormComponent.ts
@@ -4,7 +4,7 @@ import nunjucks from "nunjucks";
 import { FormComponent } from "./FormComponent";
 import { ComponentCollection } from "./ComponentCollection";
 import { FormData, FormSubmissionErrors, FormSubmissionState } from "../types";
-import { FormModel } from "../formModel";
+import { FormModel } from "../models";
 
 const getSchemaKeys = Symbol("getSchemaKeys");
