+++ runner/src/server/plugins/engine/components/DateTimeField.ts
@@ -4,7 +4,7 @@ import { InputFieldsComponents } from "@xgovformbuilder/model";
 import * as helpers from "./helpers";
 import { FormComponent } from "./FormComponent";
 import { FormData, FormSubmissionErrors, FormSubmissionState } from "../types";
-import { FormModel } from "../formModel";
+import { FormModel } from "../models";
 import { addClassOptionIfNone } from "./helpers";
 
 export class DateTimeField extends FormComponent {
