+++ runner/src/server/plugins/engine/components/TextField.ts
@@ -3,7 +3,7 @@ import { InputFieldsComponents } from "@xgovformbuilder/model";
 import * as helpers from "./helpers";
 import { FormComponent } from "./FormComponent";
 import { FormData, FormSubmissionErrors } from "../types";
-import { FormModel } from "../formModel";
+import { FormModel } from "../models";
 import { addClassOptionIfNone } from "./helpers";
 
 export class TextField extends FormComponent {
