+++ runner/src/server/plugins/engine/components/UkAddressField.ts
@@ -6,7 +6,7 @@ import * as helpers from "./helpers";
 import { FormComponent } from "./FormComponent";
 import { ComponentCollection } from "./ComponentCollection";
 import { FormData, FormSubmissionErrors, FormSubmissionState } from "../types";
-import { FormModel } from "../formModel";
+import { FormModel } from "../models";
 
 export class UkAddressField extends FormComponent {
   formChildren: ComponentCollection;
