+++ runner/src/server/plugins/engine/index.ts
@@ -1,7 +1,7 @@
-export { RelativeUrl, FeedbackContextInfo, decode } from "./feedback";
-export { findByPostcode } from "./services/addressService";
-export { configureEnginePlugin } from "./configureEnginePlugin";
+export {
+  RelativeUrl,
+  FeedbackContextInfo,
+  decodeFeedbackContextInfo,
+} from "./feedback";
 export { redirectTo, redirectUrl } from "./helpers";
-export { FormModel } from "./formModel";
-export { PageControllerBase } from "./pageControllers/pageControllerBase";
-export { plugin } from "./plugin";
+export { configureEnginePlugin } from "./configureEnginePlugin";
