+++ runner/test/cases/server/plugins/engine/feedback-context-info.test.ts
@@ -2,7 +2,7 @@ import * as Code from "@hapi/code";
 import * as Lab from "@hapi/lab";
 import {
   FeedbackContextInfo,
-  decode,
+  decodeFeedbackContextInfo,
 } from "src/server/plugins/engine/feedback";
 
 const lab = Lab.script();
@@ -14,7 +14,7 @@ suite("Feedback context info", () => {
   test("Should be able to be serialised and deserialised", () => {
     const original = new FeedbackContextInfo("My form", "My page", "/badger");
 
-    expect(decode(original.toString())).to.equal(original);
+    expect(decodeFeedbackContextInfo(original.toString())).to.equal(original);
   });
 
   test("toString should be url friendly", () => {
